#26 - refactored devices package to have separate abstract classes for cards and hubs
