+++ src/example_scripts/connect_to_star_hub.py
@@ -1,23 +1,28 @@
 from spectrumdevice.settings import CardType
-from spectrumdevice import MockSpectrumStarHub, MockSpectrumCard, SpectrumCard, SpectrumStarHub
+from spectrumdevice import (
+    MockSpectrumStarHub,
+    MockSpectrumDigitiserCard,
+    SpectrumDigitiserCard,
+    SpectrumDigitiserStarHub,
+)
 
 
-def star_hub_example(mock_mode: bool, num_cards: int, master_card_index: int) -> SpectrumStarHub:
+def star_hub_example(mock_mode: bool, num_cards: int, master_card_index: int) -> SpectrumDigitiserStarHub:
 
     if not mock_mode:
         DEVICE_IP_ADDRESS = "169.254.142.75"
         child_cards = []
         for n in range(num_cards):
             # Connect to each card in the hub.
-            child_cards.append(SpectrumCard(device_number=n, ip_address=DEVICE_IP_ADDRESS))
+            child_cards.append(SpectrumDigitiserCard(device_number=n, ip_address=DEVICE_IP_ADDRESS))
         # Connect to the hub itself
-        return SpectrumStarHub(device_number=0, child_cards=child_cards, master_card_index=master_card_index)
+        return SpectrumDigitiserStarHub(device_number=0, child_cards=child_cards, master_card_index=master_card_index)
     else:
         mock_child_cards = []
         for n in range(num_cards):
-            # Create a mock device for each card in the hub
+            # Create a mock abstract_device for each card in the hub
             mock_child_cards.append(
-                MockSpectrumCard(
+                MockSpectrumDigitiserCard(
                     device_number=n,
                     card_type=CardType.TYP_M2P5966_X4,
                     mock_source_frame_rate_hz=10.0,  # Mock devices need to be provided with a mock source frame rate
