+++ src/spectrumdevice/settings/channel.py
@@ -1,5 +1,5 @@
 """Provides an Enum defining the possible channel names assigned by the Spectrum API, and List lookup-tables for
-accessing the commands used to set vertical range and offset of each channel of a device."""
+accessing the commands used to set vertical range and offset of each channel of a abstract_device."""
 
 # Christian Baker, King's College London
 # Copyright (c) 2021 School of Biomedical Engineering & Imaging Sciences, King's College London
