+++ src/spectrumdevice/settings/device_modes.py
@@ -1,4 +1,4 @@
-"""Provides Enums defining the possible acquisition and clock modes of a spectrum device."""
+"""Provides Enums defining the possible acquisition and clock modes of a spectrum abstract_device."""
 
 # Christian Baker, King's College London
 # Copyright (c) 2021 School of Biomedical Engineering & Imaging Sciences, King's College London
