+++ src/tests/configuration.py
@@ -11,14 +11,14 @@ class SpectrumTestMode(Enum):
 
 
 # Set to TestMode.MOCK_HARDWARE for software-only testing, even if Spectrum drivers are found on the system
-# Set to TestMode.REAL_HARDWARE to run tests on a real hardware device as configured below.
+# Set to TestMode.REAL_HARDWARE to run tests on a real hardware abstract_device as configured below.
 SINGLE_CARD_TEST_MODE = SpectrumTestMode.MOCK_HARDWARE
 STAR_HUB_TEST_MODE = SpectrumTestMode.MOCK_HARDWARE
 
-# Set IP address of real spectrum device (for use if TestMode.REAL_HARDWARE is set above). Set to None to run tests on
+# Set IP address of real spectrum abstract_device (for use if TestMode.REAL_HARDWARE is set above). Set to None to run tests on
 # a local (PCIe) card.
 TEST_DEVICE_IP = "169.254.142.75"
-# Set the device number to connect to for real hardware tests. If using a star hub (e.g. netbox), this should be the
+# Set the abstract_device number to connect to for real hardware tests. If using a star hub (e.g. netbox), this should be the
 # STAR_HUB_MASTER_CARD_INDEX. If you only have a single, local (PCIe) card, set to 0.
 TEST_DEVICE_NUMBER = 1
