+++ src/tests/test_star_hub.py
@@ -1,26 +1,20 @@
 import pytest
 from numpy import array
 
-from spectrumdevice.devices.spectrum_channel import SpectrumChannel
-from spectrumdevice.devices.spectrum_star_hub import SpectrumStarHub
+from spectrum_gmbh.regs import SPC_CHENABLE
+from spectrumdevice import SpectrumDigitiserChannel, SpectrumDigitiserStarHub
+from spectrumdevice.exceptions import SpectrumInvalidNumberOfEnabledChannels
 from spectrumdevice.settings.channel import SpectrumChannelName
 from spectrumdevice.settings.transfer_buffer import CardToPCDataTransferBuffer
-from spectrumdevice.exceptions import SpectrumInvalidNumberOfEnabledChannels
+from tests.configuration import ACQUISITION_LENGTH, NUM_CARDS_IN_STAR_HUB, NUM_CHANNELS_PER_MODULE, NUM_MODULES_PER_CARD
 from tests.device_factories import create_spectrum_star_hub_for_testing
 from tests.test_single_card import SingleCardTest
-from tests.configuration import (
-    NUM_CHANNELS_PER_MODULE,
-    NUM_MODULES_PER_CARD,
-    NUM_CARDS_IN_STAR_HUB,
-    ACQUISITION_LENGTH,
-)
-from spectrum_gmbh.regs import SPC_CHENABLE
 
 
 @pytest.mark.star_hub
 class StarHubTest(SingleCardTest):
     def setUp(self) -> None:
-        self._device: SpectrumStarHub = create_spectrum_star_hub_for_testing()
+        self._device: SpectrumDigitiserStarHub = create_spectrum_star_hub_for_testing()
 
         self._expected_num_channels_each_card = NUM_CHANNELS_PER_MODULE * NUM_MODULES_PER_CARD
         self._expected_total_num_channels = self._expected_num_channels_each_card * NUM_CARDS_IN_STAR_HUB
@@ -57,7 +51,8 @@ class StarHubTest(SingleCardTest):
         expected_channels = []
         for n in range(NUM_CARDS_IN_STAR_HUB):
             expected_channels += [
-                SpectrumChannel(i, self._device._child_cards[n]) for i in range(self._expected_num_channels_each_card)
+                SpectrumDigitiserChannel(i, self._device._child_cards[n])
+                for i in range(self._expected_num_channels_each_card)
             ]
         expected_channels_tuple = tuple(expected_channels)
         self.assertEqual(expected_channels_tuple, channels)
